refactor: combine check threshold and send alert functio into prober class
